fix: 화면 공유 로직 컴포저블로 분리
