+++ client/src/components/screen/window/screenWindow.tsx
@@ -1,4 +1,4 @@
-import React from "react";
+import React, { MutableRefObject } from "react";
 import { useAtomValue } from "jotai";
 import { screenShareAtom } from "@/jotai/atom";
 import Screen from "@/components/screen/screen";
@@ -6,16 +6,16 @@ import style from "./screenWindow.module.scss";
 import { IPeer } from "@/types/peer.type";
 
 interface IProps {
-    peerList: Array<IPeer>
+    peerList: MutableRefObject<Array<IPeer>>
 }
 
 const ScreenWindow = ({ peerList }: IProps) => {
     const isScreenShare = useAtomValue(screenShareAtom);
-    const gridCols = peerList.length > 4 ? "over" : "under";
+    const gridCols = peerList.current.length > 4 ? "over" : "under";
 
     return (
         <div className={`${style.screen_window} ${style[gridCols]}`}>
-            {isScreenShare && peerList.filter(peer => peer.stream).map((peer, index) => {
+            {isScreenShare && peerList.current.filter(peer => peer.stream).map((peer, index) => {
                 if(peer.stream) {
                     return <Screen currentStream={peer.stream} key={index} userName={peer.user.userName} />;
                 }
